Created 3 new methods called setupComboBox(), setupTypePanels, and updateImage in the Pokemon Panel.
The setupComboBox() method creates a model view controller in the panel and sets it so it can be used. In Swift you do this via the GUI and interface however Java cannot do this so we must set it up this way through the code.
The setupTypePanels are set to establish the different aspects of the type panels. So far we have declared the sizes for the panels.