Added the updateTypePanels() method as well as adding code to the setupListeners() to make the pokedexDropdown to work.
pokedexDrowdown as an ActionListener is designed to let you switch between the pokemon through a dropdown menu. Meanwhile, the updateTypePanels() method lets the background change acording to type(s) including if a pokemon has multiple types.
The pokedexDropdown in the listener currently has an error with updatePokedexInfo on line 264. This is due to updatePokedexInfo not being created yet.