Created a new class called FileController to create and write into a file.
This file will currently take the information from Pokemon and create a file that has information about all the pokemon and put it as a text document. There is also a catch that will detect if there is an issue with finding the file and print out an error message if so.